Change CHAT_ID type Int -> Long

+++ src/main/kotlin/io/usoamic/swapbackend/model/Bot.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class Bot (
     @SerializedName("username") val USERNAME: String,
     @SerializedName("token") val TOKEN: String,
-    @SerializedName("chatId") val CHAT_ID: Int
+    @SerializedName("chatId") val CHAT_ID: Long
 )
